New furnitures added, randomized loot bag size and adjusted loot generation limits.
By Albert R.

Changes:
- Included graphical image assets for raid furniture
- Included said assets into loot generation and item handling methods
- Changed all weight variables from int to float to allow decimal point use
- Added new weights for furniture assets in the unity scene
- Added new serialized fields for unity scene and included the furniture images and their associated weights to said fields within unity scene
- Included photonview component with view id 2 to LootTracking class, with it an rpc call for randomizing bag size to all by master client only

+++ Assets/Raid/Scripts/Raid_InventoryItem.cs
@@ -10,7 +10,7 @@ using Photon.Pun;
 public class Raid_InventoryItem : MonoBehaviour, IPointerClickHandler
 {
     [SerializeField] private Image ItemImage;
-    [SerializeField] public int ItemWeight;
+    [SerializeField] public float ItemWeight;
     // [SerializeField] private TMP_Text ItemWeightText;
 
     public event Action<Raid_InventoryItem> OnItemClicked;
@@ -25,7 +25,7 @@ public class Raid_InventoryItem : MonoBehaviour, IPointerClickHandler
         empty = true;
     }
 
-    public void SetData(Sprite ItemSprite, int LootItemWeight)
+    public void SetData(Sprite ItemSprite, float LootItemWeight)
     {
         this.ItemImage.gameObject.SetActive(true);
         this.ItemImage.sprite = ItemSprite;
